Move components into lower dir
