+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSection/index.js
