+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSectionHeader/DerivedEntitiesSectionHeader.jsx
