+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSectionHeader/index.js
