+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSectionWrapper/DerivedEntitiesSectionWrapper.jsx
