+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSectionWrapper/index.js
