+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesSectionWrapper/style.js
