+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTable/DerivedEntitiesTable.jsx
