+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTable/index.js
