+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTable/style.js
