+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTable/utils.js
