+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTabs/index.js
