+++ context/app/static/js/components/Detail/derivedEntities/DerivedEntitiesTabs/style.js
