+++ context/app/static/js/components/Detail/derivedEntities/DerivedSamplesTable/index.js
